+++ core/contracts/oracle/implementation/VoteTiming.sol
@@ -8,7 +8,7 @@ import "../interfaces/VotingInterface.sol";
  * @title Library to compute rounds and phases for an equal length commit-reveal voting cycle.
  */
 library VoteTiming {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     struct Data {
         uint256 phaseLength;
@@ -32,7 +32,7 @@ library VoteTiming {
      * @return roundId defined as a function of the currentTime and `phaseLength` from `data`.
      */
     function computeCurrentRoundId(Data storage data, uint256 currentTime) internal view returns (uint256) {
-        uint256 roundLength = data.phaseLength.mul(uint(VotingInterface.Phase.NUM_PHASES_PLACEHOLDER));
+        uint256 roundLength = data.phaseLength.mul(uint256(VotingInterface.Phase.NUM_PHASES_PLACEHOLDER));
         return currentTime.div(roundLength);
     }
 
@@ -43,7 +43,7 @@ library VoteTiming {
      * @return timestamp unix time of when the current round will end.
      */
     function computeRoundEndTime(Data storage data, uint256 roundId) internal view returns (uint256) {
-        uint256 roundLength = data.phaseLength.mul(uint(VotingInterface.Phase.NUM_PHASES_PLACEHOLDER));
+        uint256 roundLength = data.phaseLength.mul(uint256(VotingInterface.Phase.NUM_PHASES_PLACEHOLDER));
         return roundLength.mul(roundId.add(1));
     }
 
@@ -57,7 +57,7 @@ library VoteTiming {
         // This employs some hacky casting. We could make this an if-statement if we're worried about type safety.
         return
             VotingInterface.Phase(
-                currentTime.div(data.phaseLength).mod(uint(VotingInterface.Phase.NUM_PHASES_PLACEHOLDER))
+                currentTime.div(data.phaseLength).mod(uint256(VotingInterface.Phase.NUM_PHASES_PLACEHOLDER))
             );
     }
 }
