+++ core/contracts/tokenized-derivative/LeveragedReturnCalculator.sol
@@ -25,8 +25,8 @@ contract LeveragedReturnCalculator is ReturnCalculatorInterface, Withdrawable {
     constructor(int256 _leverageMultiplier) public {
         require(_leverageMultiplier != 0);
         leverageMultiplier = _leverageMultiplier;
-        _createExclusiveRole(uint(Roles.Governance), uint(Roles.Governance), msg.sender);
-        _createWithdrawRole(uint(Roles.Withdraw), uint(Roles.Governance), msg.sender);
+        _createExclusiveRole(uint256(Roles.Governance), uint256(Roles.Governance), msg.sender);
+        _createWithdrawRole(uint256(Roles.Withdraw), uint256(Roles.Governance), msg.sender);
     }
 
     function computeReturn(int256 oldPrice, int256 newPrice) external override view returns (int256 assetReturn) {
